did-service/requirements.txt
# DID Service Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
redis>=5.0.0
cryptography>=41.0.0
pyjwt>=2.8.0

# Development dependencies
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx>=0.24.0
